refactor(advanced-search): split the semantic command module
